+++ rest-api/Core/DTOs/CoinApi/ApiResponse.cs
@@ -3,5 +3,5 @@
 public class ApiResponse<T>
 {
     public T Data { get; set; } = default!;
-    public long Timestamp { get; set; } = default;
+    public decimal Timestamp { get; set; } = default;
 }
